requirements.txt
gliner==0.2.26
torch>=2.1.0
transformers>=4.38.0
datasets>=2.18.0
huggingface_hub>=0.23.0
accelerate>=0.28.0
openai>=1.40.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0
scikit-learn>=1.3.0
seqeval>=1.2.2
